Relocated LogEntry,LogEntryEventType, LoggerExtensions into correct namespace. Removal of incorrectly copied comment.
