+++ src/DtronixPackage/ViewModel/PackageManagerViewModel.cs
@@ -308,7 +308,6 @@ namespace DtronixPackage.ViewModel
             {
                 if (result.Result == PackageOpenResultType.IncompatibleVersion)
                 {
-                    // Try opening read-only?
                     var message = new PackageMessageEventArgs(
                         PackageMessageEventArgs.MessageType.OK,
                         $"Can not open file.\r\n\r\nOpened file version is {result.OpenVersion} while application is version {openFile.AppVersion} ",
